🐛fix: 제출한 유저 수에 따른 roomStatus 변경 로직 Progress로 이관
- Fetcher은 조건부 렌더링되기 때문에 해당 로직이 동작하지 않음
